+++ application/views/article.php
@@ -9,45 +9,45 @@
               <div class="uk-card-header">
                 <div class="uk-grid uk-grid-small" data-uk-grid>
                   <div class="uk-width-expand@s">
-                    <h5 class="uk-h5 uk-text-bold"><i class="fas fa-newspaper"></i> <?= $this->news_model->getNewTitle($idlink); ?></h5>
+                    <h5 class="uk-h5 uk-text-bold"><i class="fas fa-newspaper"></i> <?= $news->title; ?></h5>
                   </div>
                   <div class="uk-width-auto@s">
-                    <p class="uk-text-small"><i class="far fa-clock"></i> <?= date('F j, Y, h:i a', $this->news_model->getNewlogDate($idlink)); ?></p>
+                    <p class="uk-text-small"><i class="far fa-clock"></i> <?= date('F j, Y, h:i a', $news->date); ?></p>
                   </div>
                 </div>
               </div>
               <div class="uk-card-body">
                 <article class="uk-article">
-                  <?= $this->news_model->getNewDescription($idlink); ?>
+                  <?= $news->description; ?>
                 </article>
               </div>
             </div>
             <div class="uk-grid uk-grid-small uk-child-width-1-1 uk-margin" data-uk-grid>
-              <?php foreach($this->news_model->getComments($idlink)->result() as $commentss): ?>
+              <?php foreach ($comments as $comment): ?>
               <div>
                 <div class="uk-card uk-card-default uk-card-body">
                   <div class="uk-grid uk-grid-small" data-uk-grid>
                     <div class="uk-width-1-6@s">
-                      <div class="Author <?php if($this->website->getRank($commentss->author) > 0) echo 'topic-author-staff'; ?> uk-flex uk-flex-center">
+                      <div class="Author <?php if($this->website->getRank($comment->author) > 0) echo 'topic-author-staff'; ?> uk-flex uk-flex-center">
                         <div class="topic-author-avatar profile">
-                          <?php if($this->base->getUserInfoGeneral($commentss->author)->num_rows()): ?>
-                          <img src="<?= base_url('assets/images/profiles/'.$this->website->getNameAvatar($this->website->getImageProfile($commentss->author))); ?>" alt="" />
+                          <?php if ($this->base->getUserInfoGeneral($comment->author)->num_rows()): ?>
+                          <img src="<?= base_url('assets/images/profiles/'.$this->website->getNameAvatar($this->website->getImageProfile($comment->author))); ?>" alt="" />
                           <?php else: ?>
                           <img src="<?= base_url('assets/images/profiles/default.png'); ?>" alt="" />
                           <?php endif; ?>
                         </div>
                       </div>
-                      <p class="uk-margin-remove uk-text-bold uk-text-center"><?= $this->website->getUsernameID($commentss->author); ?></p>
-						<?php if($this->website->getRank($commentss->author) > 0): ?>
-						<div class="author-rank-staff"><i class="fas fa-fire"></i> Staff</div>
-						<?php endif; ?>
+                      <p class="uk-margin-remove uk-text-bold uk-text-center"><?= $this->website->getUsernameID($comment->author); ?></p>
+                      <?php if ($this->website->getRank($comment->author) > 0): ?>
+                      <div class="author-rank-staff"><i class="fas fa-fire"></i> Staff</div>
+                      <?php endif; ?>
                     </div>
                     <div class="uk-width-expand@s">
-                      <p class="uk-text-small uk-text-meta uk-margin-small"><?= date('F d Y - H:i A', $commentss->date); ?></p>
-                      <?= $commentss->commentary ?>
-                      <?php if($this->website->getRank($this->session->userdata('id')) > 0 || $this->session->userdata('id') == $commentss->author && now() < strtotime('+30 minutes', $commentss->date)): ?>
+                      <p class="uk-text-small uk-text-meta uk-margin-small"><?= date('F d Y - H:i A', $comment->date); ?></p>
+                      <?= $comment->commentary ?>
+                      <?php if ($this->website->getRank($this->session->userdata('id')) > 0 || $this->session->userdata('id') == $comment->author && now() < strtotime('+30 minutes', $comment->date)): ?>
                       <div class="uk-margin-small-top">
-                        <button class="uk-button uk-button-danger uk-button-small" value="<?= $commentss->id ?>" id="button_delete<?= $commentss->id ?>" onclick="DeleteReply(event, this.value)"><i class="fas fa-eraser"></i> <?= lang('button_remove'); ?></button>
+                        <button class="uk-button uk-button-danger uk-button-small" value="<?= $comment->id ?>" id="button_delete<?= $comment->id ?>" onclick="DeleteReply(event, this.value)"><i class="fas fa-eraser"></i> <?= lang('button_remove'); ?></button>
                       </div>
                       <?php endif; ?>
                     </div>
@@ -55,33 +55,21 @@
                 </div>
               </div>
               <?php endforeach; ?>
-              <?php if(!$this->website->isLogged() && $this->forum_model->getTopicLocked($idlink) == 0): ?>
-              <div>
-                <div class="uk-card uk-card-default uk-card-body">
-                  <h3 class="uk-h3 uk-text-center"><span uk-icon="icon: comment; ratio: 1.5"></span> <?= lang('forum_comment_header'); ?></h3>
-                  <div class="glass-box-container">
-                    <p class="uk-margin-small"><?= lang('forum_comment_locked'); ?></p>
-                    <a href="<?= base_url('login'); ?>" class="uk-button uk-button-default uk-width-1-2 uk-width-1-3@m"><i class="fas fa-sign-in-alt"></i> <?= lang('button_login'); ?></a>
-                  </div>
-                </div>
-              </div>
-              <?php endif; ?>
-              <?php if($this->website->isLogged()): ?>
-              <div>
-                <div class="uk-card uk-card-default uk-card-body">
-                  <h3 class="uk-h3 uk-text-center"><span uk-icon="icon: comment; ratio: 1.5"></span> <?= lang('forum_comment_header'); ?></h3>
-                  <?= form_open('', 'id="replyForm" onsubmit="ReplyForm(event)"'); ?>
-                  <div class="uk-margin-small uk-light">
-                    <textarea class="uk-textarea tinyeditor" id="reply_comment" rows="10"></textarea>
-                  </div>
-                  <div class="uk-margin-small">
-                    <button class="uk-button uk-button-default uk-width-1-1" type="submit" id="button_reply"><i class="fas fa-reply"></i> <?= lang('button_add_reply'); ?></button>
-                  </div>
-                  <?= form_close(); ?>
-                </div>
-              </div>
-              <?php endif; ?>
             </div>
+            <?= $links; ?>
+            <?php if ($this->website->isLogged()): ?>
+            <div class="uk-card uk-card-default uk-card-body uk-margin-small">
+              <h3 class="uk-h3 uk-text-center"><span uk-icon="icon: comment; ratio: 1.5"></span> <?= lang('forum_comment_header'); ?></h3>
+              <?= form_open('', 'id="replyForm" onsubmit="ReplyForm(event)"'); ?>
+              <div class="uk-margin-small uk-light">
+                <textarea class="uk-textarea tinyeditor" id="reply_comment" rows="10"></textarea>
+              </div>
+              <div class="uk-margin-small">
+                <button class="uk-button uk-button-default uk-width-1-1" type="submit" id="button_reply"><i class="fas fa-reply"></i> <?= lang('button_add_reply'); ?></button>
+              </div>
+              <?= form_close(); ?>
+            </div>
+            <?php endif; ?>
           </div>
           <div class="uk-width-1-4@m">
             <div class="uk-card uk-card-default">
@@ -90,9 +78,9 @@
               </div>
               <div class="uk-card-body">
                 <ul class="uk-list uk-list-divider uk-text-small">
-                  <?php foreach($this->news_model->getExtendedNewsList()->result() as $list): ?>
+                  <?php foreach ($aside as $item): ?>
                   <li>
-                    <a href="<?= base_url('news/'.$list->id) ?>"><i class="far fa-newspaper"></i> <?= $list->title ?></a>
+                    <a href="<?= site_url('news/'.$item->id); ?>"><i class="far fa-newspaper"></i> <?= $item->title; ?></a>
                   </li>
                   <?php endforeach; ?>
                 </ul>
@@ -102,13 +90,13 @@
         </div>
       </div>
     </section>
-    <?php if($this->website->isLogged()): ?>
+    <?php if ($this->website->isLogged()): ?>
     <?= $tiny ?>
     <script>
       function ReplyForm(e) {
         e.preventDefault();
 
-        var news = "<?= $idlink ?>";
+        var news = "<?= $news->id; ?>";
         var reply = tinymce.get('reply_comment').getContent();
         var content = tinymce.get('reply_comment').getContent({format: 'text'}).replace('&nbsp;','').trim();
         if(content == "" || content == null || content == '<p> </p>'){
@@ -167,7 +155,7 @@
               });
             }
             $('#replyForm')[0].reset();
-            window.location.replace("<?= base_url('news/'.$idlink); ?>");
+            window.location.replace("<?= base_url('news/'.$news->id); ?>");
           }
         });
       }
@@ -213,7 +201,7 @@
                 'outEffect': 'slideRight'
               });
             }
-            window.location.replace("<?= base_url('news/'.$idlink); ?>");
+            window.location.replace("<?= base_url('news/'.$news->id); ?>");
           }
         });
       }